# backend/docs/swagger.yaml
definitions:
  modules.CryptBindingModel:
    properties:
      str:
        example: password
        type: string
    required:
    - str
    type: object
  modules.ErrorViewModel:
    properties:
      error:
        example: Error message
        type: string
    type: object
  modules.ViewObjectIdBindingModel:
    properties:
      objectId:
        example: 5ffd59f403762e429ea59a72
        type: string
    required:
    - objectId
    type: object
  modules.ViewObjectIdViewModel:
    properties:
      counter:
        type: integer
      machine:
        items:
          type: integer
        type: array
      pid:
        type: integer
      time:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/tool/crypt:
    post:
      consumes:
      - application/json
      description: 使用bcrypt加密串，返回hash值
      parameters:
      - description: 要Hash的串
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/modules.CryptBindingModel'
      produces:
      - application/json
      responses:
        "200":
          description: Hash值
          schema:
            type: string
        "400":
          description: 需要提供要Hash的串
          schema:
            $ref: '#/definitions/modules.ErrorViewModel'
      summary: 获取Crypt Hash
      tags:
      - Tool
  /api/v1/tool/uuid:
    get:
      consumes:
      - application/json
      description: 获取10个UUID，返回UUID列表
      produces:
      - application/json
      responses:
        "200":
          description: UUID列表
          schema:
            items:
              type: string
            type: array
      summary: 获取UUID
      tags:
      - Tool
  /api/v1/tool/viewObjectId:
    post:
      consumes:
      - application/json
      description: 查看ObjectId信息，包括时间戳、MAC、PID和Counter
      parameters:
      - description: 要查看的ObjectId
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/modules.ViewObjectIdBindingModel'
      produces:
      - application/json
      responses:
        "200":
          description: ObjectId信息
          schema:
            $ref: '#/definitions/modules.ViewObjectIdViewModel'
        "400":
          description: 需要提供有效的ObjectId
          schema:
            $ref: '#/definitions/modules.ErrorViewModel'
      summary: 查看ObjectId信息
      tags:
      - Tool
swagger: "2.0"